+++ packages/g-mobile-canvas/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @antv/g-mobile-canvas
 
+## 0.11.5
+
+### Patch Changes
+
+-   Updated dependencies [e8490c95]
+    -   @antv/g-plugin-canvas-renderer@1.9.14
+    -   @antv/g-plugin-canvas-picker@1.10.14
+
 ## 0.11.4
 
 ### Patch Changes